BaddieState.hitPoints is s8, not u8: retype and drop 15 no-op sign casts
Census of every retail load of BaddieState+0x354 across the cluster (engine/25,
202, 203, 206, 208, 213, 480, 578): 25 loads, and not one reads the field
zero-extended. 22 are `lbz` immediately followed by `extsb`; the other three
(Kaldachom.c 0x1194, DIM_Boss.c 0x27d0, engine/25 0xbf8) feed an `addi -1` or
`subf` whose result is truncated straight back into a `stb`, where any extension
is provably dead and MWCC drops it. There is no `cmplwi`, no `andi`, no
widening and no argument pass anywhere that takes the raw byte. A u8 field would
have to be read unsigned somewhere; this one never is.

That also settles the worry that a retype would cost an `extsb` at the plain
`hitPoints -= 1` sites: measured, it does not - MWCC elides the redundant
extension ahead of the byte store. The retype rebuilds all 30 dependent units
with zero change to any unit's fuzzy score (99.75305 / 885 complete, main.dol
OK), so the ~20 `(s8)` casts and `*(s8*)&` puns were decompilation scaffolding
restating the field's real type, not load-bearing source.

Dropped the 15 now-no-op casts in 202, 203, 206, 208, 213, 480 and 578. The
seven in engine/25/25.c are left for that unit's current owner; on an s8 field
they are ordinary no-op casts and fall to the existing no-op-cast sweep.

+++ src/dlls/objects/202/202.c
@@ -7158,7 +7158,7 @@ int iceBaddie_stateHandlerB02(GameObject* obj, GroundBaddieState* state) {
 int iceBaddie_stateHandlerB01(GameObject* obj, GroundBaddieState* state) {
     GroundBaddieState* objectState = obj->extra;
 
-    if ((s8)state->baddie.hitPoints < 1) {
+    if (state->baddie.hitPoints < 1) {
         return 3;
     }
     if ((s8)state->baddie.moveDone != 0) {
@@ -7852,7 +7852,7 @@ void iceBaddie_updateTargetCollision(GameObject* obj, int stateAddress, GroundBa
             ((IceBaddieControl*)controlAddress)->consecutiveHitCount = 0;
         }
         ((IceBaddieControl*)controlAddress)->hitTimer = 0.0f;
-        if ((s8)state->baddie.hitPoints > 0 && ((IceBaddieControl*)controlAddress)->consecutiveHitCount >= 2) {
+        if (state->baddie.hitPoints > 0 && ((IceBaddieControl*)controlAddress)->consecutiveHitCount >= 2) {
             (*gPlayerInterface)->setState(obj, state, 3);
             ((IceBaddieControl*)controlAddress)->consecutiveHitCount = 0;
             state->baddie.substate = 5;
